+++ CHANGELOG.md
@@ -10,6 +10,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Add dark theme.
 - Add clock scales.
 - Add hour marks.
+- Make clock hands translucent.
 
 ### Changed
 - Change the clock style.