+++ CHANGELOG.md
@@ -5,6 +5,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+
+## [0.0.4] - 2022-05-15
 ### Added
 - Make clock hands move smoothly.
 - Add dark theme.
@@ -31,7 +33,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Add file `.editorconfig`.
 - Add prototype.
 
-[Unreleased]: https://github.com/sakkke/clock/compare/v0.0.3...HEAD
+[Unreleased]: https://github.com/sakkke/clock/compare/v0.0.4...HEAD
+[0.0.4]: https://github.com/sakkke/clock/compare/v0.0.3...v0.0.4
 [0.0.3]: https://github.com/sakkke/clock/compare/v0.0.2...v0.0.3
 [0.0.2]: https://github.com/sakkke/clock/compare/v0.0.1...v0.0.2
 [0.0.1]: https://github.com/sakkke/clock/releases/tag/v0.0.1